requirements.txt
Django~=5.1.4
djangorestframework~=3.15.2
pip~=24.3.1
pillow~=11.0.0
django-filter~=24.3